% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{deparse}
\alias{deparse}
\title{A \code{deparse} replacement that always returns a length-1 vector}
\usage{
deparse(...)
}
\arguments{
\item{...}{Arguments passed on to \code{base::deparse()}}
}
\value{
The deparsed string
}
\description{
A \code{deparse} replacement that always returns a length-1 vector
}
\examples{
long_formula <- this_is_a_formula_with_long_variables ~
  the_test_is_checking_if_deparse_will_return +
  multiple_strings_or_not
simputation:::deparse(long_formula)
}
