% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foretell.R
\name{foretell}
\alias{foretell}
\alias{foretell.default}
\alias{foretell.glm}
\alias{foretell.rpart}
\title{Alternative to 'predict' returning values of correct type.}
\usage{
foretell(object, ...)

\method{foretell}{default}(object, ...)

\method{foretell}{glm}(object, newdata = NULL, type, ...)

\method{foretell}{rpart}(object, newdata, type, ...)
}
\arguments{
\item{object}{A model object,(\code{lm}, \code{glm}, ...)}

\item{...}{Furher arguments passed to \code{predict}.}

\item{newdata}{\code{[data.frame]} a data frame in which to look
for variables with which to predict. For \code{glmnet} models
this argument is mandatory.}

\item{type}{\code{[character]} Type of output. If missing, the type of
predicted variable is returned.}
}
\description{
Te default \code{precict} function doesn't always return the
predicted variable by default. For example, when estimating
a binomial model using \code{\link[stats]{glm}}, by default the
log-odds are returned. \code{foretell} wraps \code{predict} while
setting options so that the actual predicted value is returned.
}
