% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{na.roughfix}
\alias{na.roughfix}
\title{Rough imputation for handling missing predictors.}
\usage{
na.roughfix(object, ...)
}
\arguments{
\item{object}{an R object caryying data (e.g. \code{data.frame})}

\item{...}{arguments to be passed to other methods.}
}
\description{
This function is re-exported from
\code{\link[randomForest:na.roughfix]{randomForest:na.roughfix}} when
available. Otherwise it will throw a warning and resort to
\code{options("na.action")}
}
