% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.simputation.capabilities}
\alias{print.simputation.capabilities}
\title{print output of simputation_capabilities}
\usage{
\method{print}{simputation.capabilities}(x, ...)
}
\arguments{
\item{x}{an R object}

\item{...}{unused}
}
\description{
print output of simputation_capabilities
}
\keyword{internal}

