% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simputation-pkg.R
\docType{package}
\name{simputation-package}
\alias{simputation}
\alias{simputation-package}
\title{simputation}
\description{
A package to make imputation simpler.
}
\details{
To get started, see the \href{../doc/intro.html}{introductory vignette}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/markvanderloo/simputation}
  \item Report bugs at \url{https://github.com/markvanderloo/simputation/issues}
}

}
\author{
\strong{Maintainer}: Mark van der Loo \email{mark.vanderloo@gmail.com}

}
