% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simputation-pkg.R
\docType{package}
\name{simputation}
\alias{simputation}
\title{simputation}
\description{
A package to make imputation simpler.
}
\details{
To get started, see the \href{../doc/intro.html}{introductory vignette}.
}
