% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{simputation_capabilities}
\alias{simputation_capabilities}
\alias{simputation_suggests}
\title{Capabilities depending on suggested packages.}
\usage{
simputation_capabilities()

simputation_suggests(lib.loc = NULL)
}
\arguments{
\item{lib.loc}{Where to check whether a package is installed (passed to 
\code{\link[utils]{installed.packages}})}
}
\value{
For \code{simputation_capabilities} A named \code{character} vector of class
\code{simputation.capabilities}. The class attribute allows pretty-printing
of the output.

For \code{simputation_suggests} a \code{logical} vector, stating which
suggested packages are currently installed (\code{TRUE}) or not
(\code{FALSE}).
}
\description{
Simputation relies on a number of packages for model estimation
and/or imputation. Not all of these packages are installed when
\pkg{simputation} is installed.
}
\section{details}{


\code{simputation_capabilities} Calls every \code{impute_} function and
grabs the warning message (if any) stating that a package is missing.

\code{simputation_suggests} checks which of the suggested packages
implementing statistical models are available.
}

